+++ website/src/stateManagement/watchers/meals.js
@@ -1,11 +1,4 @@
-import { takeLeading } from 'redux-saga/effects';
-import {
-  handleDeleteMealIngredient,
-  handleGetMeals,
-  handlePatchMealIngredient,
-  handlePostMealIngredient,
-  handlePostPresetMeal,
-} from '../handlers/meals';
+import { takeLeading, call, put } from 'redux-saga/effects';
 import {
   addIngredient,
   modifyIngredient,
@@ -13,6 +6,7 @@ import {
   setMeals,
   addPresetMeal,
 } from '../reducers/meals';
+import { deleteReq, get, patch, post, postPresetMeal } from '../requests/meals';
 
 export default function* mealWatcher() {
   yield takeLeading(setMeals.type, handleGetMeals);
@@ -20,4 +14,61 @@ export default function* mealWatcher() {
   yield takeLeading(modifyIngredient.type, handlePatchMealIngredient);
   yield takeLeading(addIngredient.type, handlePostMealIngredient);
   yield takeLeading(addPresetMeal.type, handlePostPresetMeal);
+}
+
+function* handlePostMealIngredient({ payload }) {
+  try {
+    const response = yield call(post, payload);
+    const { data } = response;
+    yield put(
+      addIngredient({
+        ...payload,
+        _id: data._id,
+        ingredient: {
+          _id: data.ingredient._id,
+          ...payload.ingredient,
+        },
+      })
+    );
+  } catch (err) {
+    console.error(err.response);
+  }
+}
+
+function* handlePostPresetMeal({ payload }) {
+  try {
+    const response = yield call(postPresetMeal, payload);
+    const { data } = response;
+    yield put(addPresetMeal(data));
+  } catch (err) {
+    console.error(err.response);
+  }
+}
+
+function* handleGetMeals({ payload }) {
+  try {
+    const response = yield call(get);
+    const { data } = response;
+    yield put(setMeals(data.meals));
+  } catch (err) {
+    console.error(err.response);
+  }
+}
+
+function* handleDeleteMealIngredient({ payload }) {
+  try {
+    yield call(deleteReq, payload);
+    yield put(removeIngredient(payload));
+  } catch (err) {
+    console.error(err.response);
+  }
+}
+
+function* handlePatchMealIngredient({ payload }) {
+  try {
+    yield call(patch, payload);
+    yield put(modifyIngredient(payload));
+  } catch (err) {
+    console.error(err.response);
+  }
 }
